requirements.txt
blinker==1.6.2
click==8.1.7
dnspython==2.4.2
exceptiongroup==1.1.3
Flask==2.3.3
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.4
packaging==23.1
pandas==2.0.3
pkg_resources==0.0.0
pluggy==1.2.0
pymongo==4.5.0
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
tomli==2.0.1
tzdata==2023.3
-e git+ssh://git@github.com/laskowp1/website-project.git@615b3cd4fcb59fa0ce3611c3df8808c0c7a08e82#egg=website_project
Werkzeug==2.3.7
zipp==3.16.2
